// features/api/model/storeTankRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { StoreTankRequestCalibrationChartItem } from './storeTankRequestCalibrationChartItem';

export interface StoreTankRequest {
  /** @maxLength 255 */
  name: string;
  station_id: string;
  product_id: string;
  /** @minimum 1 */
  capacity_liters: number;
  /**
   * @minimum 0
   * @nullable
   */
  current_volume?: number | null;
  /**
   * Calibration Chart Validation
   * @nullable
   */
  calibration_chart?: StoreTankRequestCalibrationChartItem[] | null;
}